Refactor AbstractRefactoringTests class for better maintainability of test code
