+++ src/test/java/de/refactoringbot/refactorings/AbstractRefactoringTests.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Rule;
 import org.junit.rules.TemporaryFolder;
-
+import org.springframework.util.ClassUtils;
 
 /**
  * Abstract class for refactoring test cases
@@ -20,19 +20,28 @@ public abstract class AbstractRefactoringTests {
 	public TemporaryFolder folder = new TemporaryFolder();
 
 	/**
-	 * Returns a temporary copy of the file with the given fileName in the resources folder
-	 * @param fileName
+	 * Returns the test data class for a specific refactoring test class
+	 * 
+	 * @return
+	 */
+	public abstract Class<?> getTestResourcesClass();
+
+	/**
+	 * Returns a temporary copy of the test resources file of the current
+	 * refactoring test class
+	 * 
 	 * @return
 	 * @throws IOException
 	 */
-	protected File getTempCopyOfResourcesFile(String fileName) throws IOException {
+	protected File getTempCopyOfTestResourcesFile() throws IOException {
 		File copy = folder.newFile();
-		com.google.common.io.Files.copy(new File(getTestResourcesAbsolutePath() + "/" + fileName), copy);
+		com.google.common.io.Files.copy(getTestResourcesFile(), copy);
 		return copy;
 	}
-	
+
 	/**
 	 * Returns the line from the given file, stripped from whitespace
+	 * 
 	 * @param file
 	 * @param line
 	 * @return
@@ -41,18 +50,21 @@ public abstract class AbstractRefactoringTests {
 	protected String getStrippedContentFromFile(File file, int line) throws IOException {
 		String result;
 		try (Stream<String> lines = Files.lines(Paths.get(file.getAbsolutePath()))) {
-		    result = lines.skip(line-1).findFirst().get();
+			result = lines.skip(line - 1).findFirst().get();
 		}
 		return StringUtils.strip(result);
 	}
-	
+
 	/**
-	 * Returns the absolute path of the test resources folder
+	 * Returns the test resource file that was determined based on the test
+	 * resources class
+	 * 
 	 * @return
 	 */
-	private String getTestResourcesAbsolutePath() {
-		File resourcesDirectory = new File("src/test/java/de/refactoringbot/resources");
-		return resourcesDirectory.getAbsolutePath();
-	}	
+	private File getTestResourcesFile() {
+		String pathToTestResources = "src/test/java/"
+				+ ClassUtils.convertClassNameToResourcePath(getTestResourcesClass().getName()) + ".java";
+		return new File(pathToTestResources);
+	}
 
 }
