+++ src/test/java/de/refactoringbot/resources/removeparameter/TestDataClassRemoveParameter.java
@@ -1,4 +1,4 @@
-package de.refactoringbot.resources;
+package de.refactoringbot.resources.removeparameter;
 
 public class TestDataClassRemoveParameter {
